% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_holiday.R
\name{is_holiday}
\alias{is_holiday}
\title{Check if date input is an Indonesia national holiday or not}
\usage{
is_holiday(x)
}
\arguments{
\item{x}{date input}
}
\value{
TRUE or FALSE
}
\description{
Check if date input is an Indonesia national holiday or not
}
\examples{
dt <- as.Date(c("2016-01-01","2016-01-02","2016-01-03"))
is_holiday(dt)

}
