% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_holiday.R
\name{is_libur}
\alias{is_libur}
\title{Check if date input is an Indonesia public national holiday or not}
\usage{
is_libur(x, label = FALSE)
}
\arguments{
\item{x}{date input}
}
\value{
TRUE or FALSE; or return holiday description
}
\description{
Check if date input is an Indonesia public national holiday or not
}
\examples{
dt <- as.Date(c("2016-01-01","2016-01-02","2016-01-03"))
is_libur(dt)

}
