% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_weekend.R
\name{is_weekend}
\alias{is_weekend}
\alias{is_weekday}
\title{Check if date input is a weekend or not}
\usage{
is_weekend(x, wday = 1:5)

is_weekday(x, wday = 1:5)
}
\arguments{
\item{x}{date input}

\item{wday}{weekday index}
}
\value{
TRUE or FALSE
}
\description{
Check if date input is a weekend or not
}
\examples{
dt <- as.Date(c("2016-01-01","2016-01-02","2016-01-03"))
is_weekend(dt)
is_weekday(dt)
}
