% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{libur}
\alias{libur}
\title{Public Holiday in Indonesia}
\format{A tibble containing list of public holiday in Indonesia with date column}
\source{
\href{http://www.liburnasional.com/}{http://www.liburnasional.com/}
}
\usage{
libur
}
\description{
List of public holiday in Indonesia from 2013 to 2018
}
\keyword{datasets}
